requirements/constraints.txt
# requirements/constraints.txt
# Mosty:
#  • (DevOps ↔ Ekonomika podderzhki): fiksirovannye versii umenshayut «dreyf» zavisimostey i tsenu pochinki.
# Skrytye mosty:
#  • (CI ↔ Pre-commit) — odinakovye versii na lokali i v CI isklyuchayut «rabotaet u menya».
#  • (Bezopasnost ↔ Povtoryaemost) — vosproizvodimye bildy uproschayut audit CVE.
# Zemnoy abzats: constraints — eto «kalibr» dlya gayki: vse zakruchivaetsya rovno, bez lyufta po versiyam.
# c=a+b

# Core web & utils
flask==3.0.0
flask-cors==5.0.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
requests==2.32.0
werkzeug==3.0.3
ruamel.yaml==0.18.5
PyYAML==6.0.1
gunicorn==22.0.0
python-json-logger==2.0.7

# Docs & parsers
pypdf==4.2.0
pdfminer.six==20220524
python-docx==1.1.0
docx2txt==0.8
python-pptx==0.6.21
openpyxl==3.1.2
xlrd==1.2.0
ebooklib==0.18
beautifulsoup4==4.12.3
mutagen==1.47.0

# NLP
sentence-transformers==3.0.0

# Security
PyNaCl==1.5.0

# Sprint0 extras (metrics/tracing/limits/cloud)
cryptography==43.0.1
Flask-Limiter==3.7.0
prometheus-client==0.20.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-flask==0.48b0
PyJWT==2.9.0
boto3==1.35.24
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
